refactor: 워크플로우 어댑터 책임 분리
