+++ src/entities/workflow/index.ts
@@ -1,2 +1,3 @@
 export * from "./api";
+export * from "./lib";
 export * from "./model";
