+++ src/entities/workflow/lib/index.ts
@@ -1,0 +1,1 @@
+export * from "./workflow-node-adapter";
