+++ src/features/workflow-editor/model/index.ts
@@ -1,2 +1,3 @@
+export * from "./workflow-editor-adapter";
 export * from "./workflowStore";
 export * from "./useSaveWorkflowMutation";
