+++ src/shared/libs/index.ts
@@ -3,4 +3,3 @@ export * from "./auth-redirect";
 export * from "./dual-panel-layout";
 export * from "./graph";
 export * from "./query-client";
-export * from "./workflow-adapter";
